+++ apps/hops-EventStore/src/main/kotlin/no/nav/helse/hops/domain/EventStoreRepository.kt
@@ -1,6 +1,5 @@
 package no.nav.helse.hops.domain
 
-import java.time.LocalDateTime
 import java.util.UUID
 
 interface EventStoreReadOnlyRepository {
@@ -17,15 +16,3 @@ interface EventStoreReadOnlyRepository {
 interface EventStoreRepository : EventStoreReadOnlyRepository {
     suspend fun add(event: EventDto)
 }
-
-class EventDto(
-    val bundleId: UUID,
-    val messageId: UUID,
-    val eventType: String,
-    val bundleTimestamp: LocalDateTime,
-    val recorded: LocalDateTime,
-    val source: String,
-    val destinations: List<String>,
-    val data: ByteArray,
-    val dataType: String
-)
