+++ apps/hops-EventStore/src/main/kotlin/no/nav/helse/hops/domain/FhirMessageProcessService.kt
@@ -2,16 +2,10 @@ package no.nav.helse.hops.domain
 
 import ca.uhn.fhir.rest.server.exceptions.ResourceVersionConflictException
 import ca.uhn.fhir.rest.server.exceptions.UnprocessableEntityException
-import io.ktor.http.withCharset
-import no.nav.helse.hops.convert.ContentTypes
-import no.nav.helse.hops.fhir.fullyQualifiedEventType
 import no.nav.helse.hops.fhir.idAsUUID
-import no.nav.helse.hops.fhir.toJsonByteArray
-import no.nav.helse.hops.toLocalDateTime
 import no.nav.helse.hops.toUri
 import org.hl7.fhir.r4.model.Bundle
 import org.hl7.fhir.r4.model.MessageHeader
-import java.time.LocalDateTime
 import java.util.Date
 import java.util.UUID
 
@@ -19,7 +13,7 @@ class FhirMessageProcessService(private val eventStore: EventStoreRepository) {
     suspend fun process(message: Bundle) {
         validate(message)
 
-        val event = createEventDto(message)
+        val event = EventDto.create(message)
 
         eventStore.getByIdOrNull(event.messageId)?.let { existing ->
             if (event.data.contentEquals(existing.data)) return
@@ -28,21 +22,6 @@ class FhirMessageProcessService(private val eventStore: EventStoreRepository) {
 
         eventStore.add(event)
     }
-
-    private fun createEventDto(message: Bundle) =
-        (message.entry[0].resource as MessageHeader).let { header ->
-            EventDto(
-                bundleId = message.idAsUUID(),
-                messageId = header.idAsUUID(),
-                eventType = header.fullyQualifiedEventType,
-                bundleTimestamp = message.timestamp.toLocalDateTime(),
-                recorded = LocalDateTime.now(),
-                source = header.source.endpoint,
-                destinations = header.destination.map { it.endpoint }.filter { it.isNotBlank() },
-                data = message.toJsonByteArray(),
-                dataType = ContentTypes.fhirJsonR4.withCharset(Charsets.UTF_8).toString()
-            )
-        }
 
     private suspend fun validate(message: Bundle) {
         fhirValidator.validateWithResult(message).run {
